Restructure llm perf scripts to support vlm/diffusion/flux collections
