+++ scripts/performance/argument_parser.py
@@ -16,7 +16,7 @@ import argparse
 
 from nemo_run.config import NEMORUN_HOME
 
-from utils import DEFAULT_NEMO_HOME
+from .utils import DEFAULT_NEMO_HOME
 
 
 def parse_cli_args():
