+++ scripts/performance/llm/finetune_llama3_70b.py
@@ -15,13 +15,22 @@
 from os.path import basename, splitext
 
 import nemo_run as run
-from argument_parser import parse_cli_args
-from utils import get_user_configs, hf_tokenizer, set_primary_perf_configs, slurm_executor
+from ..argument_parser import parse_cli_args
+from ..utils import (
+    get_user_configs,
+    hf_tokenizer,
+    import_ckpt_experiment,
+    isfile_train_pack_metadata,
+    set_primary_perf_configs,
+    slurm_executor,
+)
 
-from nemo.collections.llm.recipes.nemotron3_22b import pretrain_recipe
+from nemo.collections.llm.gpt.data.squad import SquadDataModule
+from nemo.collections.llm.recipes.llama3_70b import finetune_recipe, model
 from nemo.collections.llm.recipes.precision.mixed_precision import bf16_with_fp8_mixed
-from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
+
+HF_MODEL_URI = "meta-llama/Meta-Llama-3-70B"
 
 
 def override_recipe_configs(
@@ -36,11 +45,19 @@ def override_recipe_configs(
     ep_size: int,
 ):
     """
-    nemotron3 22b pre-train recipe aimed at achieving best possible performance.
+    llama3 70b pre-train recipe aimed at achieving best possible performance.
 
     NOTE: Use fp8 precision training with caution. It might not give desirable results.
     """
-    recipe = pretrain_recipe(performance_mode=True)
+    finetuning_scheme = "none" if args.finetuning == "sft" else args.finetuning
+    gpu_type = args.gpu.lower()
+    if gpu_type in ["gb200"] and finetuning_scheme == "lora":
+        # On GB200 for lora task, we need to enable Cuda Graph for optimal performance.
+        # However, Cuda Graph increases memory usage, so in order to avoid OOM, we need
+        # to reduce the sequence length.
+        recipe = finetune_recipe(peft_scheme=finetuning_scheme, performance_mode=True, seq_length=2048)
+    else:
+        recipe = finetune_recipe(peft_scheme=finetuning_scheme, performance_mode=True)
     recipe = set_primary_perf_configs(
         recipe,
         args.tensorboard,
@@ -57,19 +74,20 @@ def override_recipe_configs(
     )
 
     # data module configs
-    recipe.data.num_train_samples = args.max_steps * gbs * mbs  # ensure only 1 epoch for whole run
-    if args.compute_dtype == "bf16" or args.gpu.lower() == "b200":
-        recipe.data.tokenizer = run.Config(
-            get_nmt_tokenizer, library="null", model_name="NullTokenizer", vocab_size=256000
-        )
-        recipe.model.tokenizer = recipe.data.tokenizer
-    else:
-        recipe.data.tokenizer = hf_tokenizer("nvidia/megatron-gpt2-345m")
+    recipe.data.tokenizer = hf_tokenizer(HF_MODEL_URI)
+    if recipe.data.__fn_or_cls__ == SquadDataModule and not isfile_train_pack_metadata(HF_MODEL_URI, recipe.data):
+        # flag is valid only for SquadDataModule
+        recipe.data.force_redownload = True
 
     # compute dtype configs
     if args.compute_dtype.lower() == "fp8":
         recipe.trainer.plugins = bf16_with_fp8_mixed()
         recipe.trainer.plugins.grad_reduce_in_fp32 = False
+
+    enable_cuda_graph = bool(args.gpu.lower() in ["gb200"] and finetuning_scheme == "lora")
+    recipe.model.config.enable_cuda_graph = enable_cuda_graph
+    recipe.trainer.strategy.use_te_rng_tracker = enable_cuda_graph
+    recipe.data.packed_sequence_specs.pad_cu_seqlens = enable_cuda_graph
 
     return recipe
 
@@ -77,13 +95,13 @@ def override_recipe_configs(
 if __name__ == "__main__":
     args = parse_cli_args().parse_args()
 
-    kwargs = get_user_configs(args.gpu.lower(), "pre_train", "nemotron3", "22b", args)
+    kwargs = get_user_configs(args.gpu.lower(), args.finetuning, "llama3", "70b", args)
     num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size, _ = kwargs
 
     recipe = override_recipe_configs(args, num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size)
 
     exp_config = f"{num_nodes}nodes_tp{tp_size}_pp{pp_size}_cp{cp_size}_vp{vp_size}_{mbs}mbs_{gbs}gbs"
-    exp_name = f"{splitext(basename(__file__))[0]}_{args.compute_dtype}_{exp_config}"
+    exp_name = f"{args.finetuning}_{splitext(basename(__file__))[0]}_{args.compute_dtype}_{exp_config}"
 
     executor = slurm_executor(
         args.account,
@@ -104,6 +122,7 @@ if __name__ == "__main__":
         plugins.append(NsysPlugin(start_step=5, end_step=6))
 
     with run.Experiment(exp_name) as exp:
+        exp.add(*import_ckpt_experiment(executor, model(), source=f"hf://{HF_MODEL_URI}"))
         exp.add(
             recipe,
             executor=executor,
