+++ scripts/performance/llm/pretrain_nemotron3_22b.py
@@ -13,14 +13,14 @@
 # limitations under the License.
 
 from os.path import basename, splitext
-from typing import Optional
 
 import nemo_run as run
-from argument_parser import parse_cli_args
-from utils import get_user_configs, hf_tokenizer, set_primary_perf_configs, slurm_executor
+from ..argument_parser import parse_cli_args
+from ..utils import get_user_configs, hf_tokenizer, set_primary_perf_configs, slurm_executor
 
-from nemo.collections.llm.recipes.mixtral_8x7b import pretrain_recipe
+from nemo.collections.llm.recipes.nemotron3_22b import pretrain_recipe
 from nemo.collections.llm.recipes.precision.mixed_precision import bf16_with_fp8_mixed
+from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
 
 
@@ -34,10 +34,9 @@ def override_recipe_configs(
     cp_size: int,
     vp_size: int,
     ep_size: int,
-    etp_size: Optional[int],
 ):
     """
-    mixtral 8x7b pre-train recipe aimed at achieving best possible performance.
+    nemotron3 22b pre-train recipe aimed at achieving best possible performance.
 
     NOTE: Use fp8 precision training with caution. It might not give desirable results.
     """
@@ -55,21 +54,22 @@ def override_recipe_configs(
         cp_size,
         vp_size,
         ep_size,
-        etp_size,
     )
 
     # data module configs
     recipe.data.num_train_samples = args.max_steps * gbs * mbs  # ensure only 1 epoch for whole run
-    recipe.data.tokenizer = hf_tokenizer("mistralai/Mixtral-8x7B-v0.1")
+    if args.compute_dtype == "bf16" or args.gpu.lower() == "b200":
+        recipe.data.tokenizer = run.Config(
+            get_nmt_tokenizer, library="null", model_name="NullTokenizer", vocab_size=256000
+        )
+        recipe.model.tokenizer = recipe.data.tokenizer
+    else:
+        recipe.data.tokenizer = hf_tokenizer("nvidia/megatron-gpt2-345m")
 
     # compute dtype configs
     if args.compute_dtype.lower() == "fp8":
         recipe.trainer.plugins = bf16_with_fp8_mixed()
         recipe.trainer.plugins.grad_reduce_in_fp32 = False
-
-    enable_cuda_graph = bool(args.gpu.lower() in ["b200"])
-    recipe.model.config.enable_cuda_graph = enable_cuda_graph
-    recipe.trainer.strategy.use_te_rng_tracker = enable_cuda_graph
 
     return recipe
 
@@ -77,14 +77,12 @@ def override_recipe_configs(
 if __name__ == "__main__":
     args = parse_cli_args().parse_args()
 
-    kwargs = get_user_configs(args.gpu.lower(), "pre_train", "mixtral", "8x7b", args)
-    num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size, etp_size = kwargs
+    kwargs = get_user_configs(args.gpu.lower(), "pre_train", "nemotron3", "22b", args)
+    num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size, _ = kwargs
 
-    recipe = override_recipe_configs(args, num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size, etp_size)
+    recipe = override_recipe_configs(args, num_nodes, mbs, gbs, tp_size, pp_size, cp_size, vp_size, ep_size)
 
-    exp_config = (
-        f"{num_nodes}nodes_tp{tp_size}_pp{pp_size}_cp{cp_size}_vp{vp_size}_ep{ep_size}_etp{etp_size}_{mbs}mbs_{gbs}gbs"
-    )
+    exp_config = f"{num_nodes}nodes_tp{tp_size}_pp{pp_size}_cp{cp_size}_vp{vp_size}_{mbs}mbs_{gbs}gbs"
     exp_name = f"{splitext(basename(__file__))[0]}_{args.compute_dtype}_{exp_config}"
 
     executor = slurm_executor(
