+++ scripts/performance/recommended_model_configs/model_configs_h100.csv
