+++ scripts/performance/recommended_model_configs/strong_scaling_model_configs_h100.csv
